+++ R/use-rates.R
@@ -1,7 +1,7 @@
 #' Calculate application rates for active ingredients
 #'
 #' An application rate in g active substance/ha is calculated from information
-#' on dosage (product concentration in the application solution), application volume,
+#' on dosage (product concentration in the application solution) and application volume,
 #' or directly from the product application rate. This is complicated by the fact
 #' that a rate ("expenditure" in the XML file) with units l/ha can refer
 #' to the application solution or to the liquid product.
@@ -42,7 +42,16 @@
 #' library(srppp)
 #' library(dplyr, warn.conflicts = FALSE)
 #' library(dm, warn.conflicts = FALSE)
-#' sr <- srppp_dm()
+#'
+#' sr <- try(srppp_dm())
+#'
+#' # Fall back to internal test data if downloading or reading fails
+#' if (inherits(sr, "try-error")) {
+#'   sr <- system.file("testdata/Daten_Pflanzenschutzmittelverzeichnis_2024-12-16.zip",
+#'       package = "srppp") |>
+#'     srppp_xml_get_from_path(from = "2024-12-16") |>
+#'     srppp_dm()
+#' }
 #'
 #' product_uses_with_ingredients <- sr$substances |>
 #'   filter(substance_de %in% c("Halauxifen-methyl", "Kupfer (als Kalkpr\u00E4parat)")) |>
@@ -107,11 +116,15 @@ application_rate_g_per_ha <- function(product_uses,
         if_else(is.na(source), # if no external information, assume l/ha is product
           if_else(is.na(g_per_L), # if g_per_L is not defined
             if (skip_l_per_ha_without_g_per_L) NA # as in the 2023 indicator
-            else rate * dosage * (percent/100), # assume l/ha to be water,
-            # dosage is assumed to be g product per L. This is correct for
-            # Rhodofix 2009 (Grünbuch) and 2012 (XML)
+            else {
+              if_else(is.na(dosage),
+                # If we have no dosage, treat l/ha as kg/ha and use percent to calculate rate_g_per_ha.
+                rate * (percent/100) * 1000, # Correct for Metro 2017
+                # If we have a dosage, the rate in l/ha is assumed to be the application solution
+                rate * dosage * (percent/100)) # Correct for Rhodofix 2009 (Grünbuch) and 2012 (XML)
+            },
             rate * g_per_L), # l/ha is product
-          if (fix_l_per_ha) { # Sometimes we have information that l/ha is water
+          if (fix_l_per_ha) { # Sometimes we have external information that the rate in l/ha is water
             rate * dosage/100 * g_per_L
           } else NA),
       units_de == "kg/ha" ~ rate * (percent * 10), # percent w/w means 10 g/kg