+++ NEWS.md
@@ -1,4 +1,15 @@
-## version 1.0.1
+## version 1.1.0
+
+- Example code and vignettes: To satisfy CRAN policy, fall back to use the register version distributed as test data with the package (see below) in case downloading  or reading from the standard URL fails, instead of throwing an error. This makes the check process on CRAN robust against server outage or a changed URL.
+- Include information on the completeness of the effect against each target organism in the column `type` of the `pests` table of the `srppp_dm` object. Possible types are `PEST_FULL_EFFECT`, `PEST_PARTIAL_EFFECT` and `PEST_SIDE_EFFECT`. In the `obligations` table, include a column `varying_effect`, that contains `TRUE` in case the obligation text indicates that a full effect has not been demonstrated in all cases.
+- `R/resolve_cultures.R`: Function to expand a dataframe with culture names at the lowest level (leafs) of a culture tree constructed by `srppp_dm`.
+- Add a culture tree as a `Node` object from the `data.tree` package as attribute to `srppp_dm` objects and show it in the main vignette.
+- Remove the vignette intended for JOSS, as it was rejected there
+- Remove duplicate ingredient entries for the same substance in the same product (issue #6)
+- `inst/testdata/`: Add the zipped XML from 16 December 2024 for testing purposes and test the resulting `srppp_dm` object for referential integrity
+- `srppp_dm()`: Add argument `verbose` with default `TRUE`.
+
+## version 1.0.1 (CRAN release: 2024-11-01)
 
 - Trim leading and trailing whitespace from descriptions that are read in using the internal function `get_descriptions()`, addressing an issue Elisabeth raised in srppphist
 - Remove html documentation built with pkgdown from the git repository, the online documentation is built using the github workflow